Added selective asset bundle building
- New B&S Asset Builder window
- Moved Build AssetBundles button to the new window
- Added a way to select specific asset bundles for building, so you do not build all of them each time

+++ BasAssetsCreator/Assets/SDK/Scripts/Editor/EditorTools.cs
@@ -1,25 +1,110 @@
 ﻿using UnityEditor;
 using System.IO;
+using System;
 using UnityEngine;
 using System.Collections.Generic;
-
-public class EditorTools
+public class EditorTools : EditorWindow
 {
-    [MenuItem("Blade and Sorcery/Build AssetBundles")]
-    static void BuildAllAssetBundles()
+    public Dictionary<string, bool> exportBundle = new Dictionary<string, bool>();
+    List<string> bundleNames;
+    List<AssetBundleBuild> assetBundleBuilds;
+    [MenuItem("Blade and Sorcery/Asset Builder")] 
+    public static void ShowWindow()
     {
+        EditorWindow.GetWindow<EditorTools>("B&S Asset Builder");
+    }
+    
+    private void OnFocus()  
+    {
+        bundleNames = new List<string>(AssetDatabase.GetAllAssetBundleNames());
+        foreach (string bundle in bundleNames)
+        {
+            try {
+                exportBundle[bundle] = EditorPrefs.GetBool("ExportBundle" + bundle);
+                } catch (Exception) 
+            {
+                exportBundle[bundle] = true;
+                EditorPrefs.SetBool("ExportBundle" + bundle, true);
+            }
+        }
+    }
+     
+    private void OnGUI()
+    {
+        GUILayout.Space(10);
+        GUILayout.Label(new GUIContent("  Select the following assets to export"), new GUIStyle("BoldLabel"));
+        GUILayout.Space(10);
+
+        foreach (string bundle in bundleNames)
+        {
+            exportBundle[bundle] = GUILayout.Toggle(exportBundle[bundle], bundle);
+        }
+
+        GUILayout.Space(10);
+
+        Handles.BeginGUI();
+        Handles.color = Color.black;
+        Handles.DrawLine(new Vector3(0, 36), new Vector3(10000, 36));
+        Handles.EndGUI();
+
+        if (GUILayout.Button("Build Asset Bundles"))
+        {
+            BuildAllAssetBundles();
+        }
+
+    }
+
+    private void OnDisable()
+    {
+        foreach (string bundle in bundleNames)
+        {
+            try { 
+            EditorPrefs.SetBool("ExportBundle" + bundle, exportBundle[bundle]);
+            } catch (Exception)
+            {
+                EditorPrefs.SetBool("ExportBundle" + bundle, true);
+            }
+        }
+    }
+
+    private void BuildAllAssetBundles()
+    {
+        assetBundleBuilds = new List<AssetBundleBuild>();
         string assetBundleDirectory = "Assets/AssetBundles";
         DirectoryInfo dir = new DirectoryInfo(assetBundleDirectory);
         if (!Directory.Exists(assetBundleDirectory)) Directory.CreateDirectory(assetBundleDirectory);
-
-        foreach (FileInfo file in dir.GetFiles("*.*"))
+        
+        foreach (string bundle in bundleNames)
         {
-            file.Delete();
+            foreach (FileInfo file in dir.GetFiles("*.*"))
+            {
+                if (file.Extension != ".bundle" || exportBundle[bundle])
+                {
+                    file.Delete();
+                }
+            }
+            if (exportBundle[bundle]) { 
+            AssetBundleBuild build = new AssetBundleBuild
+            {
+                assetBundleName = bundle,
+                assetBundleVariant = AssetImporter.GetAtPath(assetBundleDirectory).assetBundleVariant,
+                assetNames = AssetDatabase.GetAssetPathsFromAssetBundle(bundle)
+            };
+                assetBundleBuilds.Add(build);
+            }
+            try
+            {
+                EditorPrefs.SetBool("ExportBundle" + bundle, exportBundle[bundle]);
+            }
+            catch (Exception)
+            {
+                EditorPrefs.SetBool("ExportBundle" + bundle, true);
+            }
         }
 
-        BuildPipeline.BuildAssetBundles(assetBundleDirectory, BuildAssetBundleOptions.None, BuildTarget.StandaloneWindows);
+        BuildPipeline.BuildAssetBundles(assetBundleDirectory, assetBundleBuilds.ToArray(), BuildAssetBundleOptions.None, BuildTarget.StandaloneWindows);
 
-        List<string> bundleNames = new List<string>(AssetDatabase.GetAllAssetBundleNames());
+        bundleNames = new List<string>(AssetDatabase.GetAllAssetBundleNames());
 
         foreach (FileInfo file in dir.GetFiles("*.*"))
         {
@@ -27,12 +112,20 @@ public class EditorTools
             {
                 file.MoveTo(file.FullName + ".bundle");
             }
-            else
+            else if (file.Extension != ".bundle")
             {
                 file.Delete();
             }
         }
+
         AssetDatabase.Refresh();
-        Debug.Log("Asset bundles created in " + dir.FullName);
+        if (assetBundleBuilds.ToArray().Length > 0) {
+            string bundleNames = "";
+            foreach (AssetBundleBuild build in assetBundleBuilds)
+            {
+                bundleNames = bundleNames + build.assetBundleName + ", ";
+            }
+        Debug.Log("Created Asset Bundles " + bundleNames + "in " + dir.FullName);
+        }
     }
 }
